refactor: create global type definitions for common models
